+++ src/main/java/pl/asie/computronics/integration/waila/ConfigValues.java
@@ -1,5 +1,7 @@
 package pl.asie.computronics.integration.waila;
 
+import cpw.mods.fml.common.Loader;
+import cpw.mods.fml.common.ModAPIManager;
 import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaRegistrar;
 import pl.asie.computronics.reference.Mods;
@@ -9,7 +11,8 @@ import pl.asie.computronics.reference.Mods;
  */
 public enum ConfigValues {
 
-	Address(Mods.Computronics + ".enableAddress"),
+	OCAddress(Mods.Computronics + ".enableOCAddress", Mods.OpenComputers),
+	NCAddress(Mods.Computronics + ".enableNCAddress", Mods.NedoComputers),
 	Tape(Mods.Computronics + ".enableTape"),
 	TapeName(Mods.Computronics + ".enableTapeName"),
 	DriveState(Mods.Computronics + ".enableDriveState"),
@@ -18,6 +21,7 @@ public enum ConfigValues {
 
 	private String key;
 	private boolean defvalue;
+	private String modID;
 
 	private ConfigValues(String key) {
 		this(key, true);
@@ -28,17 +32,30 @@ public enum ConfigValues {
 		this.defvalue = defvalue;
 	}
 
+	private ConfigValues(String key, String modID) {
+		this(key, modID, true);
+	}
+
+	private ConfigValues(String key, String modID, boolean defvalue) {
+		this.key = key;
+		this.defvalue = defvalue;
+		this.modID = modID;
+	}
+
 	private void registerConfigRemote(IWailaRegistrar reg) {
 		reg.addConfigRemote(Mods.Computronics_NAME, key, defvalue);
 	}
 
 	static void registerConfigs(IWailaRegistrar reg) {
 		for(ConfigValues value : ConfigValues.values()) {
-			value.registerConfigRemote(reg);
+			if(value.modID == null || Loader.isModLoaded(value.modID) || ModAPIManager.INSTANCE.hasAPI(value.modID)) {
+				value.registerConfigRemote(reg);
+			}
 		}
 	}
 
 	public boolean getValue(IWailaConfigHandler config) {
-		return config.getConfig(key, defvalue);
+		return (this.modID == null || Loader.isModLoaded(this.modID) || ModAPIManager.INSTANCE.hasAPI(this.modID))
+			&& config.getConfig(key, defvalue);
 	}
 }