+++ src/main/java/pl/asie/computronics/integration/waila/WailaComputronics.java
@@ -4,6 +4,9 @@ import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
 import mcp.mobius.waila.api.IWailaDataProvider;
 import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.world.World;
 
 import java.util.List;
 
@@ -55,5 +58,17 @@ public class WailaComputronics implements IWailaDataProvider {
 		}
 
 		return currenttip;
+	}
+
+	@Override
+	public NBTTagCompound getNBTData(TileEntity te, NBTTagCompound tag, World world, int x, int y, int z) {
+
+		for(WailaProviders p : WailaProviders.VALUES) {
+			if(p.isInstance(te.getBlockType())) {
+				tag = p.getProvider().getNBTData(te, tag, world, x, y, z);
+			}
+		}
+
+		return tag;
 	}
 }