+++ src/main/java/pl/asie/computronics/integration/waila/providers/WailaColorfulLamp.java
@@ -4,7 +4,10 @@ import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.world.World;
 import pl.asie.computronics.integration.waila.ConfigValues;
+import pl.asie.computronics.tile.TileColorfulLamp;
 import pl.asie.computronics.util.StringUtil;
 
 import java.util.List;
@@ -31,5 +34,13 @@ public class WailaColorfulLamp extends ComputronicsWailaProvider {
 		currenttip.add(StringUtil.localizeAndFormat("tooltip.computronics.waila.lamp.green", g));
 		currenttip.add(StringUtil.localizeAndFormat("tooltip.computronics.waila.lamp.blue", b));
 		return currenttip;
+	}
+
+	@Override
+	public NBTTagCompound getNBTData(TileEntity te, NBTTagCompound tag, World world, int x, int y, int z) {
+		if(te != null && te instanceof TileColorfulLamp) {
+			tag.setShort("clc", (short) (((TileColorfulLamp) te).getLampColor() & 32767));
+		}
+		return tag;
 	}
 }